+++ .vscode/extensions.json
@@ -1,8 +1,6 @@
 {
   "recommendations": [
     "nrwl.angular-console",
-    "angular.ng-template",
-    "esbenp.prettier-vscode",
-    "zeplin.zeplin"
+    "angular.ng-template"
   ]
 }
