+++ apps/festival/festival/src/app/dashboard/dashboard.module.ts
@@ -12,7 +12,7 @@ import { OrgNameModule } from '@blockframes/organization/pipes/org-name.pipe';
 import { ToLabelModule } from '@blockframes/utils/pipes';
 import { MovieFormShellModule } from '@blockframes/movie/form/shell/shell.module';
 import { MovieShellConfig } from '@blockframes/movie/form/movie.shell.config';
-import { FORMS_CONFIG } from '@blockframes/movie/form/shell/shell.component';
+import { FORMS_CONFIG } from '@blockframes/movie/form/movie.shell.interfaces';
 
 // Tunnel routes
 import { tunnelRoutes } from './tunnel/movie-tunnel.routes';
