+++ libs/movie/src/lib/movie/form/shell/shell.component.html
@@ -12,7 +12,16 @@
     <tunnel-nav [steps]="steps$ | async"></tunnel-nav>
   </aside>
   <ng-template #confirmExit>
-    <h1>Hey</h1>
-    <p>Test</p>
+    <div class="confirm-exit" fxLayout="column" fxLayoutAlign="start center">
+      <h1>You are going to leave the Movie Form.</h1>
+      <p>Pay attention, if you leave now your changes will not be saved.</p>
+      <div fxLayout="row" fxLayoutAlign="center">
+        <button class="save" mat-flat-button color="primary" [mat-dialog-close]="true">
+          Save & Exit</button>
+        <button class="text-primary" mat-button [mat-dialog-close]="false">
+          Close without saving
+        </button>
+      </div>
+    </div>
   </ng-template>
 </tunnel-layout>
