+++ libs/movie/src/lib/movie/form/shell/shell.component.scss
@@ -8,3 +8,7 @@ header {
     height: 48px;
   }
 }
+
+.confirm-exit {
+  padding: 16px;
+}
