+++ libs/movie/src/lib/movie/form/shell/shell.component.ts
@@ -1,20 +1,17 @@
 // Angular
-import { Component, ChangeDetectionStrategy, OnInit, Inject, AfterViewInit, OnDestroy, InjectionToken, ChangeDetectorRef, ViewChild } from '@angular/core';
+import { Component, ChangeDetectionStrategy, OnInit, Inject, AfterViewInit, OnDestroy, ChangeDetectorRef, ViewChild } from '@angular/core';
 import { DOCUMENT } from '@angular/common';
 
 // Blockframes
 import { MovieQuery } from '@blockframes/movie/+state';
 import { TunnelRoot, TunnelStep, TunnelLayoutComponent } from '@blockframes/ui/tunnel';
+import { FORMS_CONFIG, ShellConfig } from '../movie.shell.interfaces';
 
 // RxJs
 import { map, startWith } from 'rxjs/operators';
-import { Observable, of, Subscription, combineLatest } from 'rxjs';
+import { Observable, Subscription, combineLatest } from 'rxjs';
 import { ProductionStatus } from '@blockframes/utils/static-model';
-import { EntityControl, FormEntity } from '@blockframes/utils/form';
-import type { MovieShellConfig } from '../movie.shell.config';
-import type { CampaignShellConfig } from '@blockframes/campaign/form/campaign.shell.config';
 import { RouterQuery } from '@datorama/akita-ng-router-store';
-
 
 function isStatus(prodStatus: ProductionStatus, acceptableStatus: ProductionStatus[]) {
   return acceptableStatus.includes(prodStatus)
@@ -105,19 +102,6 @@ function getSteps(status: ProductionStatus, appSteps: TunnelStep[] = []): Tunnel
   })
 }
 
-export interface FormShellConfig<Control extends EntityControl<Entity>, Entity> {
-  form: FormEntity<Control, Entity>;
-  onInit(): Observable<any>[];
-  onSave(publishing: boolean): Promise<any>
-}
-
-export interface ShellConfig {
-  movie: MovieShellConfig;
-  campaign?: CampaignShellConfig
-}
-
-export const FORMS_CONFIG = new InjectionToken<ShellConfig>('List of form managed by the shell');
-
 @Component({
   selector: 'movie-form-shell',
   templateUrl: './shell.component.html',
@@ -139,7 +123,7 @@ export class MovieFormShellComponent implements TunnelRoot, OnInit, AfterViewIni
   ) { }
 
   ngOnInit() {
-    const subs: Observable<any>[] = Object.values(this.configs).map(config => config.onInit()).flat();
+    const subs: Observable<any>[] = Object.values(this.configs).map((config: any) => config.onInit()).flat();
     this.sub = combineLatest(subs).subscribe(() => this.cdr.markForCheck());
     this.exitRoute = `/c/o/dashboard/title/${this.query.getActiveId()}`;
   }
